Explicitly install vim after checking versions
It occurred to me that I want to be able to run the install script
without installing things that are irrelevant (eg .startxwinrc on
non-Cygwin systems), so rather than installing everything, test whether
it's sensible to install, and only install if it is.

For Vim, that means only installing the Vim files if Vim is installed,
and warning if the Vim version is one I haven't seen before (ie anything
other than 7.3).