Better parsing of git svn find-rev output
When parsing `git svn find-rev` output as part of the `rev-to-hash`
alias, parse line-by-line rather than token-by-token for commit hashes,
and discard anything that isn't a commit hash.

This means the command will no longer attempt to parse "rebuilding
cache" messages, which cause it to find far more commits than it ought.